Revert "Put this SierraItemRecord VHS in a shared fixture"
This reverts commit 5d9679c3e43ce2cd7fa3cb925f9c4bd24b08bfad.
